+++ Changes.md
@@ -5,6 +5,7 @@ Change Log in version 3.10.1.5 (2021022305)
 3. Fix 'My Courses dropdown not showing all enrolled courses' - #258.
 4. Fix "My courses dropdown can show hidden courses when user does not have capability when 'mysitessortoverride' setting is not
    set to 'Use list from my overview'", related to #258.
+5. Fix 'One Topic font color change not getting applied' - #264.
 
 Change Log in version 3.10.1.4 (2021022304)
 ===========================================